### 2024-05-31 #6817 @editorialbot
[PRE REVIEW]: f3dasm: Framework for Data-Driven Design and Analysis of Structures and Materials
**Submitting author:** @mpvanderschelling (<a href="http://orcid.org/0000-0003-3602-0452">Martin van der Schelling</a>)
**Repository:** https://github.com/bessagroup/f3dasm
**Branch with paper.md** (empty if default branch):
**Version:** v1.5.1
