### 2024-06-20 197133d
Update version to 1.6.0
