+++ src/f3dasm/__version__.py
@@ -1,1 +1,1 @@
-__version__: str = "1.5.1"
+__version__: str = "1.6.0"
