### 2024-06-20 8918dca
Round input data to 6 decimal places and refactor combine_data_to_multiindex function
