### 2024-06-20 6b0da8b
Update numpy version in requirements.txt
