+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy
+numpy<2.0.0
 scipy
 pandas
 xarray
