### 2024-06-20 36af2ce
Refactor column renaming methods in experimentdata module
