### 2024-06-20 #6817 @RMeli
@idocx, @AntObi thank you very much for accepting to review this submission for JOSS. I'll start the review process now, but label it as "paused" until the first of July. I'll also set an automatic reminder in 2 weeks.
