+++ src/f3dasm/_src/experimentdata/_experimental/_jobqueue2.py
@@ -4,15 +4,16 @@
 from __future__ import annotations
 
 # Standard
+from copy import deepcopy
 from enum import Enum
 from pathlib import Path
-from typing import Iterable, List, Optional, Type
+from typing import Iterable, List, Type
 
 # Third-party
 import pandas as pd
 
 # Local
-from ._data import _Data
+from ._newdata2 import _Data
 
 #                                                          Authorship & Credits
 # =============================================================================
@@ -46,68 +47,43 @@ class NoOpenJobsError(Exception):
     def __init__(self, message):
         super().__init__(message)
 
+# =============================================================================
 
-class _JobQueue:
-    def __init__(self, jobs: Optional[pd.Series] = None):
-        """
-        A class that represents a dictionary of jobs that
-        can be marked as 'open', 'in progress', finished', or 'error'.
 
-        Parameters
-        ----------
-        filename : str
-            The name of the file that the jobs will be saved in.
-        """
-        if jobs is None:
-            jobs = pd.Series(dtype='string')
+class Index:
+    def __init__(self, jobs: pd.Series | None | str = None):
+        if isinstance(jobs, str):
+            self.jobs = pd.Series(jobs, index=[0], dtype='string')
 
-        self.jobs: pd.Series = jobs
+        elif jobs is None:
+            self.jobs = pd.Series(dtype='string')
 
-    def __add__(self, other: _JobQueue | str) -> _JobQueue:
-        """Add two JobQueue objects together.
+        else:
+            self.jobs = jobs
 
-        Parameters
-        ----------
-        other : JobQueue
-            JobQueue object to add.
+    def __len__(self) -> int:
+        return len(self.jobs)
 
-        Returns
-        -------
-        JobQueue
-            JobQueue object containing the added jobs.
-        """
-        if isinstance(other, str):
-            # make _JobQueue from the jobnumber
-            other = _JobQueue(
-                pd.Series(other, index=[0], dtype='string'))
+    def __add__(self, __o: Index | str) -> Index:
+        if isinstance(__o, str):
+            __o = Index(__o)
 
-        try:
-            last_index = self.jobs.index[-1]
-        except IndexError:  # Empty Series
-            return _JobQueue(other.jobs)
+        if self.jobs.empty:
+            return __o
 
         # Make a copy of other.jobs and modify its index
-        other_jobs_copy = other.jobs.copy()
-        other_jobs_copy.index = other_jobs_copy.index + last_index + 1
-        return _JobQueue(pd.concat([self.jobs, other_jobs_copy]))
+        other_jobs_copy = deepcopy(__o)
+        other_jobs_copy.jobs.index = range(
+            len(other_jobs_copy)) + self.jobs.index[-1] + 1
 
-    def __getitem__(self, index: int | slice | Iterable[int]) -> _JobQueue:
-        """Get a subset of the data.
+        return Index(pd.concat([self.jobs, other_jobs_copy.jobs]))
 
-        Parameters
-        ----------
-        index : int, slice, list
-            The index of the data to get.
+    def __getitem__(self, indices: int | slice | Iterable[int]) -> Index:
+        if isinstance(indices, int):
+            indices = [indices]
+        return Index(self.jobs[indices].copy())
 
-        Returns
-        -------
-            A subset of the data.
-        """
-        if isinstance(index, int):
-            index = [index]
-        return _JobQueue(self.jobs[index].copy())
-
-    def __eq__(self, __o: _JobQueue) -> bool:
+    def __eq__(self, __o: Index) -> bool:
         return self.jobs.equals(__o.jobs)
 
     def _repr_html_(self) -> str:
@@ -117,12 +93,16 @@ class _JobQueue:
     def indices(self) -> pd.Index:
[…]
