+++ src/f3dasm/_src/experimentdata/_experimental/_newdata2.py
@@ -1,0 +1,221 @@
+#                                                                       Modules
+# =============================================================================
+
+from __future__ import annotations
+
+# Standard
+from copy import deepcopy
+from pathlib import Path
+from typing import Any, Dict, Iterable, List, Optional, Type, Union
+
+# Third-party
+import numpy as np
+import pandas as pd
+import xarray as xr
+
+#                                                          Authorship & Credits
+# =============================================================================
+__author__ = 'Martin van der Schelling (M.P.vanderSchelling@tudelft.nl)'
+__credits__ = ['Martin van der Schelling']
+__status__ = 'Stable'
+# =============================================================================
+#
+# =============================================================================
+
+MISSING_VALUE = np.nan
+
+
+class _Data:
+    def __init__(self, data: Dict[int, Dict[str, Any]] = None):
+        self.data = data if data is not None else {}
+
+    def __len__(self) -> int:
+        return len(self.data)
+
+    def __iter__(self):
+        return iter(self.data.values())
+
+    def __getitem__(self, rows: int | slice | Iterable[int]) -> _Data:
+
+        if isinstance(rows, int):
+            rows = [rows]
+
+        return _Data({row: self.data.get(row, {}) for row in rows})
+
+    def __add__(self, __o: _Data) -> _Data:
+        if self.is_empty():
+            return __o
+
+        _data_copy = deepcopy(self)
+        other_data_copy = deepcopy(__o)
+
+        new_indices = (np.array(range(len(__o))) + max(self.data) + 1).tolist()
+
+        _data_copy.data.update({row: values for row, values in zip(
+            new_indices, other_data_copy.data.values())})
+        return _data_copy
+
+    def __eq__(self, __o: _Data) -> bool:
+        return self.data == __o.data
+
+    def _repr_html_(self) -> str:
+        return self.to_dataframe()._repr_html_()
+
+    def __repr__(self) -> str:
+        return self.to_dataframe().__repr__()
+
+    @property
+    def indices(self) -> List[int]:
+        return list(self.data.keys())
+
+    @property
+    def names(self) -> List[str]:
+        return self.to_dataframe().columns.tolist()
+
+    @classmethod
+    def from_indices(cls, rows: Iterable[int]):
+        return cls({row: {} for row in rows})
+
+    # @classmethod
+    # def from_domain(cls, space: Iterable[str]):
+    #     return cls(None)
+
+    @classmethod
+    def from_file(cls, filename: Path) -> _Data:
+        ...
+
+    @classmethod
+    def from_numpy(cls: Type[_Data], array: np.ndarray,
+                   keys: Optional[Iterable[str]] = None) -> _Data:
+        if keys is not None:
+            return _Data(
+                {index: {key: col for key, col in zip(keys, row)
+                         } for index, row in enumerate(array)})
+        else:
+            # Look out! i is now an integer key!
+            return _Data(
+                {index: {i: col for i, col in enumerate(row)
+                         } for index, row in enumerate(array)})
+
+    @classmethod
+    def from_dataframe(cls, df: pd.DataFrame) -> _Data:
+        return _Data(
+            {index: row.to_dict() for index, (_, row) in
+             enumerate(df.iterrows())})
+
+    def to_numpy(self) -> np.ndarray:
+        return self.to_dataframe().to_numpy()
+
+    def to_xarray(self, label: str):
+        df = self.to_dataframe()
+        # Can create the xarray with the information from the domain!
+        return xr.DataArray(
+            self.to_dataframe(), dims=['iterations', label], coords={
[…]

[…]
