### 2024-06-25 9be5928
Added docstings and tests for _newdata2 object
