+++ src/f3dasm/_src/experimentdata/_experimental/_newdata2.py
@@ -24,25 +24,76 @@ __status__ = 'Stable'
 
 MISSING_VALUE = np.nan
 
+# =============================================================================
+
 
 class _Data:
     def __init__(self, data: Dict[int, Dict[str, Any]] = None):
+        """
+        Initialize the _Data object.
+
+        Parameters
+        ----------
+        data : Dict[int, Dict[str, Any]], optional
+            The data dictionary with integer keys and dictionaries as values.
+        """
         self.data = data if data is not None else {}
 
     def __len__(self) -> int:
+        """
+        Get the number of items in the data.
+
+        Returns
+        -------
+        int
+            Number of items in the data.
+        """
         return len(self.data)
 
     def __iter__(self):
+        """
+        Get an iterator over the data values.
+
+        Returns
+        -------
+        iterator
+            Iterator over the data values.
+        """
         return iter(self.data.values())
 
     def __getitem__(self, rows: int | slice | Iterable[int]) -> _Data:
+        """
+        Get a subset of the data.
 
+        Parameters
+        ----------
+        rows : int or slice or Iterable[int]
+            The rows to retrieve.
+
+        Returns
+        -------
+        _Data
+            The subset of the data.
+        """
         if isinstance(rows, int):
             rows = [rows]
 
         return _Data({row: self.data.get(row, {}) for row in rows})
 
     def __add__(self, __o: _Data) -> _Data:
+        """
+        Add another _Data object to this one.
+
+        Parameters
+        ----------
+        __o : _Data
+            The other _Data object.
+
+        Returns
+        -------
+        _Data
+            The combined _Data object.
+        """
         if self.is_empty():
             return __o
 
@@ -56,78 +107,254 @@ class _Data:
         return _data_copy
 
     def __eq__(self, __o: _Data) -> bool:
+        """
+        Check if another _Data object is equal to this one.
+
+        Parameters
+        ----------
+        __o : _Data
+            The other _Data object.
+
+        Returns
+        -------
+        bool
+            True if the objects are equal, False otherwise.
+        """
         return self.data == __o.data
 
     def _repr_html_(self) -> str:
+        """
+        Get the HTML representation of the data.
+
+        Returns
+        -------
+        str
+            The HTML representation of the data.
+        """
         return self.to_dataframe()._repr_html_()
 
     def __repr__(self) -> str:
+        """
+        Get the string representation of the data.
+
+        Returns
+        -------
+        str
+            The string representation of the data.
+        """
         return self.to_dataframe().__repr__()
+
+
+#                                                                    Properties
+# =============================================================================
 
     @property
     def indices(self) -> List[int]:
+        """
+        Get the indices of the data.
+
+        Returns
+        -------
+        List[int]
+            The list of indices.
+        """
         return list(self.data.keys())
 
     @property
     def names(self) -> List[str]:
+        """
+        Get the column names of the data.
+
+        Returns
+        -------
+        List[str]
+            The list of column names.
+        """
         return self.to_dataframe().columns.tolist()
 
-    @classmethod
-    def from_indices(cls, rows: Iterable[int]):
-        return cls({row: {} for row in rows})
+    def is_empty(self) -> bool:
+        """
+        Check if the data is empty.
 
-    # @classmethod
-    # def from_domain(cls, space: Iterable[str]):
-    #     return cls(None)
+        Returns
+        -------
+        bool
+            True if the data is empty, False otherwise.
+        """
+        return not bool(self.data)
+
+
+#                                                                Initialization
+# =============================================================================
+
+    @classmethod
+    def from_indices(cls, rows: Iterable[int]) -> _Data:
+        """
+        Create a _Data object from a list of indices.
[…]
