+++ tests/newdata/test_data.py
@@ -1,292 +1,287 @@
 from copy import deepcopy
-from typing import Any, List
+from typing import Any, Dict, List
 
 import numpy as np
 import pandas as pd
 import pytest
+import xarray as xr
 
-from f3dasm._src.experimentdata._columns import _Columns
-from f3dasm._src.experimentdata._newdata import _Data, _Index
-from f3dasm.design import Domain
+from f3dasm._src.experimentdata._experimental._newdata2 import (
+    _convert_dict_to_data, _Data, _data_factory)
 
 pytestmark = pytest.mark.smoke
 
-DataType = List[List[Any]]
+DataType = Dict[int, Dict[str, Any]]
+
+#                                                                Initialization
+# =============================================================================
 
 
-def test_init(list_1: DataType):
-    data = _Data(list_1)
-    assert data.data == list_1
-    assert data.columns.names == [0, 1, 2]
-    assert data.indices.equals(pd.Index([0, 1, 2]))
+def test_init():
+    data = _Data({0: {"a": 1, "b": 2}})
+    assert len(data) == 1
+    assert not data.is_empty()
+    assert data.data == {0: {"a": 1, "b": 2}}
 
 
-def test_init_with_columns(list_1: DataType, columns_1: _Columns):
-    data = _Data(list_1, columns_1)
-    assert data.data == list_1
-    assert data.names == ['a', 'b', 'c']
+def test_init_empty():
+    data = _Data()
+    assert len(data) == 0
+    assert data.is_empty()
 
 
-def test_init_with_columns_and_indices(
-        list_1: DataType, columns_1: _Columns, indices_1: _Index):
-    data = _Data(list_1, columns_1, indices_1)
-    assert data.data == list_1
-    assert data.names == ['a', 'b', 'c']
-    assert data.indices.equals(pd.Index([3, 5, 6]))
+def test_init_with_data():
+    input_data = {0: {"a": 1, "b": 2}}
+    data = _Data(input_data)
+    assert len(data) == 1
+    assert not data.is_empty()
+    assert data.data == input_data
 
 
-def test__len__(list_1: DataType):
-    data = _Data(list_1)
-    assert len(data) == 3
+def test_from_numpy():
+    array = np.array([[1, 2, 3], [4, 5, 6]])
+    data = _Data.from_numpy(array)
+    expected_data = {0: {0: 1, 1: 2, 2: 3}, 1: {0: 4, 1: 5, 2: 6}}
+    assert data.data == expected_data
 
 
-def test__iter__(list_1: DataType):
-    data = _Data(list_1)
-    for i, row in enumerate(data):
-        assert row == list_1[i]
+def test_from_numpy_with_keys():
+    array = np.array([[1, 2, 3], [4, 5, 6]])
+    data = _Data.from_numpy(array, keys=["a", "b", "c"])
+    expected_data = {0: {"a": 1, "b": 2, "c": 3}, 1: {"a": 4, "b": 5, "c": 6}}
+    assert data.data == expected_data
 
 
-def test__getitem__(list_1: DataType):
-    data = _Data(list_1)
-    assert data[0].data[0] == list_1[0]
-    assert data[1].data[0] == list_1[1]
-    assert data[2].data[0] == list_1[2]
+def test_from_dataframe():
+    df = pd.DataFrame({"a": [1, 2], "b": [3, 4]})
+    data = _Data.from_dataframe(df)
+    expected_data = {0: {"a": 1, "b": 3}, 1: {"a": 2, "b": 4}}
[…]
