### 2024-06-25 6d99381
Fix data indexing issue and add column renaming functionality
