+++ tests/newdata/experimentdata/test_experimentdata.py
@@ -1,0 +1,737 @@
+from __future__ import annotations
+
+import csv
+import pickle
+from pathlib import Path
+from typing import Iterable
+
+import numpy as np
+import pandas as pd
+import pytest
+import xarray as xr
+
+from f3dasm import ExperimentSample
+from f3dasm._src.design.parameter import _ContinuousParameter
+from f3dasm._src.experimentdata._experimental._jobqueue2 import \
+    Index as _JobQueue
+from f3dasm._src.experimentdata._experimental._newdata2 import DataTypes, _Data
+from f3dasm._src.experimentdata._experimental._newexperimentdata2 import \
+    ExperimentData
+from f3dasm.design import Domain, Status, make_nd_continuous_domain
+
+pytestmark = pytest.mark.smoke
+
+SEED = 42
+
+
+def test_check_experimentdata(experimentdata: ExperimentData):
+    assert isinstance(experimentdata, ExperimentData)
+
+# Write test functions
+
+
+def test_experiment_data_init(experimentdata: ExperimentData, domain: Domain):
+    assert experimentdata.domain == domain
+    assert experimentdata.project_dir == Path.cwd()
+    # Add more assertions as needed
+
+
+def test_experiment_data_add(experimentdata: ExperimentData,
+                             experimentdata2: ExperimentData, domain: Domain):
+    experimentdata_total = ExperimentData(domain)
+    experimentdata_total.add_experiments(experimentdata)
+    experimentdata_total.add_experiments(experimentdata2)
+    assert experimentdata_total == experimentdata + experimentdata2
+
+
+def test_experiment_data_len_empty(domain: Domain):
+    experiment_data = ExperimentData(domain)
+    assert len(experiment_data) == 0  # Update with the expected length
+
+
+def test_experiment_data_len_equals_input_data(experimentdata: ExperimentData):
+    assert len(experimentdata) == len(experimentdata._input_data)
+
+
+@pytest.mark.parametrize("slice_type", [3, [0, 1, 3]])
+def test_experiment_data_select(slice_type: int | Iterable[int], experimentdata: ExperimentData):
+    input_data = experimentdata._input_data[slice_type]
+    output_data = experimentdata._output_data[slice_type]
+    jobs = experimentdata._jobs[slice_type]
+    constructed_experimentdata = ExperimentData(
+        input_data=input_data, output_data=output_data, jobs=jobs, domain=experimentdata.domain)
+    assert constructed_experimentdata == experimentdata.select(slice_type)
+
+#                                                                           Constructors
+# ======================================================================================
+
+
+def test_from_file(experimentdata_continuous: ExperimentData, seed: int, tmp_path: Path):
+    # experimentdata_continuous.filename = tmp_path / 'test001'
+    experimentdata_continuous.store(tmp_path / 'experimentdata')
+
+    experimentdata_from_file = ExperimentData.from_file(
+        tmp_path / 'experimentdata')
+
+    # Check if the input_data attribute of ExperimentData matches the expected_data
+    pd.testing.assert_frame_equal(
+        experimentdata_continuous._input_data.to_dataframe(), experimentdata_from_file._input_data.to_dataframe(), check_dtype=False, atol=1e-6)
+    pd.testing.assert_frame_equal(experimentdata_continuous._output_data.to_dataframe(),
+                                  experimentdata_from_file._output_data.to_dataframe())
+    pd.testing.assert_series_equal(
[…]
