+++ tests/newdata/experimentdata/conftest.py
@@ -1,0 +1,130 @@
+from __future__ import annotations
+
+import numpy as np
+import pandas as pd
+import pytest
+import xarray as xr
+
+from f3dasm._src.design.parameter import (_CategoricalParameter,
+                                          _ContinuousParameter,
+                                          _DiscreteParameter)
+from f3dasm._src.experimentdata._experimental._newexperimentdata2 import \
+    ExperimentData
+from f3dasm.design import Domain, make_nd_continuous_domain
+
+SEED = 42
+
+
+@pytest.fixture(scope="package")
+def seed() -> int:
+    return SEED
+
+
+@pytest.fixture(scope="package")
+def domain() -> Domain:
+
+    space = {
+        'x1': _ContinuousParameter(-5.12, 5.12),
+        'x2': _DiscreteParameter(-3, 3),
+        'x3': _CategoricalParameter(["red", "green", "blue"])
+    }
+
+    return Domain(space=space)
+
+
+@pytest.fixture(scope="package")
+def domain_continuous() -> Domain:
+    return make_nd_continuous_domain(bounds=np.array([[0., 1.], [0., 1.], [0., 1.]]), dimensionality=3)
+
+
+@pytest.fixture(scope="package")
+def experimentdata(domain: Domain) -> ExperimentData:
+    e_data = ExperimentData(domain)
+    e_data.sample(sampler='random', n_samples=10, seed=SEED)
+    return e_data
+
+
+@pytest.fixture(scope="package")
+def experimentdata2(domain: Domain) -> ExperimentData:
+    return ExperimentData.from_sampling(sampler='random', domain=domain, n_samples=10, seed=SEED)
+
+
+@pytest.fixture(scope="package")
+def experimentdata_continuous(domain_continuous: Domain) -> ExperimentData:
+    return ExperimentData.from_sampling(sampler='random', domain=domain_continuous, n_samples=10, seed=SEED)
+
+
+@pytest.fixture(scope="package")
+def experimentdata_expected() -> ExperimentData:
+    domain_continuous = make_nd_continuous_domain(
+        bounds=np.array([[0., 1.], [0., 1.], [0., 1.]]), dimensionality=3)
+    data = ExperimentData.from_sampling(
+        sampler='random', domain=domain_continuous, n_samples=10, seed=SEED)
+    for es, output in zip(data, np.zeros((10, 1))):
+        es.store(name='y', object=float(output))
+        data._set_experiment_sample(es)
+    data.add(input_data=np.array([[0.0, 0.0, 0.0], [1.0, 1.0, 1.0]]),
+             output_data=np.array([[0.0], [0.0]]), domain=data.domain)
+
+    # data._input_data.data = data._input_data.data.round(6)
+    return data
+
+
+@pytest.fixture(scope="package")
+def experimentdata_expected_no_output() -> ExperimentData:
+    domain_continuous = make_nd_continuous_domain(
+        bounds=np.array([[0., 1.], [0., 1.], [0., 1.]]), dimensionality=3)
+    data = ExperimentData.from_sampling(
+        sampler='random', domain=domain_continuous, n_samples=10, seed=SEED)
+    data.add(input_data=np.array(
+        [[0.0, 0.0, 0.0], [1.0, 1.0, 1.0]]), domain=domain_continuous)
+
+    # data._input_data.data = data._input_data.data.round(6)
+
+    return data
+
+
+@pytest.fixture(scope="package")
+def experimentdata_expected_only_domain() -> ExperimentData:
+    domain_continuous = make_nd_continuous_domain(
+        bounds=np.array([[0., 1.], [0., 1.], [0., 1.]]), dimensionality=3)
[…]
