### 2024-06-25 74fd315
Refactor domain initialization and data loading
