+++ src/f3dasm/_src/experimentdata/_experimental/_jobqueue2.py
@@ -35,6 +35,8 @@ class Status(str, Enum):
     def __str__(self) -> str:
         return self.value
 
+# =============================================================================
+
 
 class NoOpenJobsError(Exception):
     """
@@ -52,6 +54,14 @@ class NoOpenJobsError(Exception):
 
 class Index:
     def __init__(self, jobs: pd.Series | None | str = None):
+        """
+        Initializes the Index object.
+
+        Parameters
+        ----------
+        jobs : pd.Series, None, or str, optional
+            Series of jobs, None, or a single job as a string.
+        """
         if isinstance(jobs, str):
             self.jobs = pd.Series(jobs, index=[0], dtype='string')
 
@@ -62,9 +72,30 @@ class Index:
             self.jobs = jobs
 
     def __len__(self) -> int:
+        """
+        Returns the number of jobs.
+
+        Returns
+        -------
+        int
+            Number of jobs.
+        """
         return len(self.jobs)
 
     def __add__(self, __o: Index | str) -> Index:
+        """
+        Adds another Index or a string to this Index.
+
+        Parameters
+        ----------
+        __o : Index or str
+            Another Index object or a string representing a job.
+
+        Returns
+        -------
+        Index
+            A new Index object containing the combined jobs.
+        """
         if isinstance(__o, str):
             __o = Index(__o)
 
@@ -73,180 +104,89 @@ class Index:
 
         # Make a copy of other.jobs and modify its index
         other_jobs_copy = deepcopy(__o)
-        other_jobs_copy.jobs.index = pd.Index(range(
-            len(other_jobs_copy))) + self.jobs.index[-1] + 1
+        other_jobs_copy.jobs.index = pd.Index(
+            range(len(other_jobs_copy))) + self.jobs.index[-1] + 1
 
         return Index(pd.concat([self.jobs, other_jobs_copy.jobs]))
 
     def __getitem__(self, indices: int | slice | Iterable[int]) -> Index:
+        """
+        Gets a subset of jobs by indices.
+
+        Parameters
+        ----------
+        indices : int, slice, or Iterable[int]
+            Indices to get.
+
+        Returns
+        -------
+        Index
+            A new Index object containing the selected jobs.
+        """
         if isinstance(indices, int):
             indices = [indices]
         return Index(self.jobs[indices].copy())
 
     def __eq__(self, __o: Index) -> bool:
+        """
+        Checks if this Index is equal to another Index.
+
+        Parameters
+        ----------
+        __o : Index
+            Another Index object to compare.
+
+        Returns
+        -------
+        bool
+            True if the two Index objects are equal, False otherwise.
+        """
         return self.jobs.equals(__o.jobs)
 
     def _repr_html_(self) -> str:
+        """
+        Returns an HTML representation of the jobs.
+
+        Returns
+        -------
+        str
+            HTML representation of the jobs.
+        """
         return self.jobs.__repr__()
 
     @property
     def indices(self) -> pd.Index:
-        """The indices of the jobs."""
+        """
+        The indices of the jobs.
+
+        Returns
+        -------
+        pd.Index
+            The indices of the jobs.
+        """
         return self.jobs.index
 
-    def iloc(self, indices: Iterable[int]) -> Iterable[int]:
+    def iloc(self, indices: Iterable[int] | int) -> Iterable[int]:
+        """
+        Gets the position of the given indices in the jobs.
+
+        Parameters
+        ----------
+        indices : Iterable[int] or int
+            Indices to locate.
+
+        Returns
+        -------
+        Iterable[int]
+            Positions of the given indices.
+        """
+        if isinstance(indices, int):
+            indices = [indices]
         return self.indices.get_indexer(indices)
 
-    #                                                  Alternative Constructors
-    # =========================================================================
-
-    @classmethod
[…]
