### 2024-06-26 1eaae8b
Added tmp file creation after storing. Fixes #273
