### 2024-06-26 0678959
Remove old files and rename files with correct extensions
