### 2024-06-26 ca4606c
Add optional parameter to store methods
