### 2024-06-26 f63e1d3
Fix reading empty Pandas DataFrame error
