### 2024-07-17 #6912 @idocx
Hey Martin! Thank you for the great work!

I will try to get the review ready by the end of this week. So far it looks good except for the annoying `numpy 2.0` issue. The Pypi registry has not been updated to `numpy < 2.0` to avoid `autograd`  package failure. (https://github.com/HIPS/autograd/issues/622#issue-2371900518)

Do you have plan to update the pypi registry some time? I notice that you have updated the version on the github. Thanks!
