### 2024-07-17 5f28b38
Updated version to 1.5.2 in order to push change to pypi registry
