### 2024-07-17 #6912 @mpvanderschelling
@idocx ; I have updated the [PyPI registry](https://pypi.org/project/f3dasm/1.5.2/) to version 1.5.2; which pushes the `numpy 2.0` issue
