### 2024-07-23 bessagroup/f3dasm#274 @mpvanderschelling
Sphinx Gallery problem with TOC headers
## Problem
In the tutorial doc part, all the tutorial pages go into the last section.

## Possible solution

This might be some header-issue when generating the Sphinx Gallery
