### 2024-07-23 #6912 @mpvanderschelling
Hey Yuxing @idocx,

Thank you for your feedback, I'll address each of your points individually

**Documentation**
Yes, this is definitely a bug from my part (created a new issue [here](https://github.com/bessagroup/f3dasm/issues/274)), and might be due to some header issues when generating the Sphinx Gallery. Will take a look at it!

**Project structure**
The reason that I put all the source code of the Python package in a separate `_src` directory is precisely what you mentioned: since I put the actual implementations behind a private submodule, I can structure the public API separately [^1]. It indeed creates overhead when implementing new public functions/classes, but I have more control over the public API and notice earlier in the development process when I (accidentally) cause an API change when renaming methods.

**Non-vectorizable parameters**
At the moment, this is not implemented yet but on the backlog. This behaviour is already implemented for [large objects or array-like data](https://f3dasm.readthedocs.io/en/latest/auto_examples/003_datageneration/003_storing.html#large-objects-and-array-like-data), and makes it possible to store any output that is non-vectorizable to disk and store a reference link in the `ExperimentData` object. I am planning on implementing the same behaviour for the input parameters in the near future.

**Data generator**
Yes, for that you will have to use the [class approach of creating a custom `DataGenerator`](https://f3dasm.readthedocs.io/en/latest/auto_examples/003_datageneration/001_own_datagenerator.html#using-the-datagenerator-class). You can create any class and store object parameters (e.g., different config for VASP) via the `self` attribute. In your `execute` function, you can use those. Whenever you call the `evaluate` method on your `ExperimentData` object, you can determine which custom `DataGenerator` object you want to use on the experiments.

[^1]: Similar practices are shown for the [numpy](https://github.com/numpy/numpy/tree/main/numpy) with the `_core` directory and [jax](https://github.com/google/jax/tree/main/jax) source code

**Database support**
Yes, there is the idea to support database access to the `ExperimentData` object, but this is not a priority at the moment. However, I have no experience with databases and Python, so if you have any suggestions or pointers for this, please let me know!

Best,
Martin
