### 2024-08-02 968dc1f
Sphinx Gallery problem with TOC headers Fixes #274
