### 2024-08-02 bessagroup/f3dasm#274 @mpvanderschelling
Fixes by adding the `sphinx_gallery_conf` option `nested_sections=False`. See [here](https://sphinx-gallery.github.io/stable/configuration.html#nesting-gallery-sections) for explanation
