### 2024-08-02 #6912 @mpvanderschelling
Hey everyone,
- @idocx ; I have resolved the [Sphing Gallery proble with TOC headers](https://github.com/bessagroup/f3dasm/issues/274) issue
- @AntObi ; do you have an estimation when you are able to finish your review :)
