### 2024-08-05 #6912 @idocx
Hi @RMeli and @mpvanderschelling, the author's response has resolved all my concerns. I agree to publish the software package to JOSS as it is.

For the database support, you can refer to [pandas.DataFrame.to_sql](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.to_sql.html) if you would like to support SQL database. If you would like to use MongoDB, then you can directly call `df.to_dict("record")` to get the record you would like to save. One tricky thing I can imagine is to retrieve large result - you will have to link the local file to the database somehow. But it should be doable.

Thanks,
Yuxing
