### 2024-08-05 #6912 @AntObi
Hi @RMeli and @mpvanderschelling , this is some great work that has been produced! It's well documented, great coverage with unit tests and the integration of experiments with Hydra is a very great addition to the package.

 I'm happy to accept the paper and package as is.

Yours truly,
Anthony
