### 2024-08-05 #6912 @RMeli
@mpvanderschelling can you please add the DOI for Optuna? I thin it should be [https://doi.org/10.1145/3292500.333070](https://doi.org/10.1145/3292500.3330701)
