+++ paper/paper.bib
@@ -53,7 +53,8 @@
 } 
 
 @inproceedings{Akiba2019,
-  doi={https://doi.org/10.1145/3292500.3330701},
+  doi={10.1145/3292500.3330701},
+  url={https://doi.org/10.1145/3292500.3330701},
   title={Optuna: A next-generation hyperparameter optimization framework},
   author={Akiba, Takuya and Sano, Shotaro and Yanase, Toshihiko and Ohta, Takeru and Koyama, Masanori},
   booktitle={Proceedings of the 25th ACM SIGKDD international conference on knowledge discovery \& data mining},
