### 2024-08-05 #6912 @mpvanderschelling
Latest release: [1.5.3](https://github.com/bessagroup/f3dasm/releases/tag/v1.5.3)
Zenodo DOI: [https://doi.org/10.5281/zenodo.13227916](https://doi.org/10.5281/zenodo.13227916)
