### 2024-08-12 #6912 @RMeli
Thank you @mpvanderschelling for the fast final fix. Based on the extensive comments from the reviewers (and my own assessment), I'm happy to recommend this paper for acceptance! The EiC will now perform the final checks. Thank you for engaging so well with the review process.

Thank you very much again @idocx and @AntObi for doing the hard work!
