### 2024-08-12 #6912 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
