### 2024-08-15 e09f934
Update version to 1.5.4
