### 2024-08-16 fbde020
remove local draft of JOSS since this has been published
