+++ paper/paper.bib
@@ -1,76 +1,0 @@
-@Article{Bessa2017,
-  author   = {Bessa, Miguel and Bostanabad, R. and Liu, Z. and Hu, A. and Apley, Daniel W. and Brinson, C. and Chen, W. and Liu, Wing Kam},
-  journal  = {Computer Methods in Applied Mechanics and Engineering},
-  title    = {{A framework for data-driven analysis of materials under uncertainty: Countering the curse of dimensionality}},
-  year     = {2017},
-  issn     = {00457825},
-  number   = {April},
-  pages    = {633--667},
-  volume   = {320},
-  abstract = {A new data-driven computational framework is developed to assist in the design and modeling of new material systems and structures. The proposed framework integrates three general steps: (1) design of experiments, where the input variables describing material geometry (microstructure), phase properties and external conditions are sampled; (2) efficient computational analyses of each design sample, leading to the creation of a material response database; and (3) machine learning applied to this database to obtain a new design or response model. In addition, the authors address the longstanding challenge of developing a data-driven approach applicable to problems that involve unacceptable computational expense when solved by standard analysis methods – e.g. finite element analysis of representative volume elements involving plasticity and damage. In these cases the framework includes the recently developed “self-consistent clustering analysis” method in order to build large databases suitable for machine learning. The authors believe that this will open new avenues to finding innovative materials with new capabilities in an era of high-throughput computing (“big-data”).},
-  doi      = {10.1016/j.cma.2017.03.037},
-  file     = {:home/martin/Documents/Mendeley Desktop/Bessa et al. - 2017 - A framework for data-driven analysis of materials under uncertainty Countering the curse of dimensionality.pdf:pdf},
-  keywords = {Design of experiments, Machine learning and data mining, Plasticity, Reduced order model, Self-consistent clustering analysis},
-}
-
-@article{Sundberg2022, 
-  doi = {10.21105/joss.04364}, 
-  url = {https://doi.org/10.21105/joss.04364}, 
-  year = {2022}, 
-  publisher = {The Open Journal}, 
-  volume = {7}, 
-  number = {75}, 
-  pages = {4364}, 
-  author = {Jack D. Sundberg and Siona S. Benjamin and Lauren M. McRae and Scott C. Warren}, 
-  title = {Simmate: a framework for materials science}, 
-  journal = {Journal of Open Source Software}
-}
-
-@article{Pietka2022, 
-  doi = {10.21105/joss.04719}, 
-  url = {https://doi.org/10.21105/joss.04719}, 
-  year = {2022}, 
-  publisher = {The Open Journal}, 
-  volume = {7}, 
-  number = {79}, 
-  pages = {4719}, 
-  author = {Isabel Pietka and Ralf Drautz and Thomas Hammerschmidt}, 
-  title = {strucscan: A lightweight Python-based framework for high-throughput material simulation}, 
-  journal = {Journal of Open Source Software}
-}
-
-@article{Biscani2020, 
-  doi = {10.21105/joss.02338}, 
-  url = {https://doi.org/10.21105/joss.02338}, 
-  year = {2020}, 
-  publisher = {The Open Journal}, 
-  volume = {5}, 
-  number = {53}, 
-  pages = {2338}, 
-  author = {Francesco Biscani and Dario Izzo}, 
-  title = {A parallel global multiobjective framework for optimization: pagmo}, 
-  journal = {Journal of Open Source Software} 
-} 
-
-@inproceedings{Akiba2019,
-  doi={10.1145/3292500.3330701},
-  url={https://doi.org/10.1145/3292500.3330701},
[…]
