### 2024-08-16 7a8e4f4
add DOI of JOSS publication
