### 2024-08-16 #6912 @mpvanderschelling
Before I leave; I'm interested in becoming a review myself for JOSS :smile: .
I have signed up for this already, but maybe good to mention here as well
