### 2024-08-16 f2b8608
fixed logo link
