### 2024-08-16 f9dbeb4
fixed logo link attempt 2
