### 2024-08-16 #6912 @RMeli
That's great! I can indeed see you in the list at https://reviewers.joss.theoj.org.

However, your "topic areas" is empty, so editors will not know what your area of expertise is. I'd strongly suggest to fill that section in your profile, with well known keywords. Editors usually look for those in order to find potential reviewers.
