### 2024-08-16 #6912 @RMeli
I think "topic areas" correspond to "domains" in your profile. I see you did fill the "Area(s) of expertise" field already.
