### 2024-08-19 #6912 @Kevin-Mattheus-Moerman
@mpvanderschelling et al. congratulations on this JOSS publication!

@RMeli thanks for editing!
